Improve feedback widget (#197)
- widget moved to the left, and is part of the html output rather than
being a floating js object. Because the right ToC has `position: sticky`
and scrolls itself as you progress through a page, I was unable to leave
the feedback on the right _and_ avoid it shadowing some content. I've
made the nav on the left slightly smaller so people can't think there's
anything underneath the feedback. It gets expanded to full height when
the feedback vanishes.
- after leaving feedback, politely thanks you and vanishes within a
couple seconds
- you _can't_ leave negative feedback without adding a message
- you _can_ leave a message to a positive feedback (but may also leave
it empty)
- the user journey through docs pages is registered and submitted as
part of their feedback
- taken away the `Edit this page` links, and improved style/wording
overall

----

Depends on 

https://github.com/neo4j-labs/neo4j-labs.github.io/pull/3
https://github.com/neo4j-contrib/dx-feedback-form/pull/2