Add beta-until label (#254)
For new and deprectaed labels it is already possible to include a
version number. Although we have a generic `beta` label that can be used
to display just the text `Beta`, there is no way to display a version
with the text, for example `Beta until 5.12`.

This PR adds a new `label--beta-until-<version>` role to be used for
beta labels where a version number is also required.