fix tab position in viewport (#284)
Updates the tab scroll to take the elements position before and after
all the tabs have been switched so that the difference can be used for
the `scrollTo`. The `scrollTo` behavior changes to instant, because we
are effectively masking the fact that the tab position would move
without a scrollTo, and we are moving it back to where it was before the
tab was clicked.