helper to choose edit this page link text (#295)
Allow different text for the 'edit this page' link.

- If the page attribute `page-edit-url-text` is set that value is used.
- If no attribute is set, the text is determined by the `page-theme`.
  - For `docs` pages, the text will be "Raise an issue".
  - For all other pages the text is "Edit this page".
- The default title is 'Docs Feedback' followed by the file path and
branch or tag name that the feedback relates to. To replace the 'Docs
Feedback' text, use the page attribute `page-edit-url-title`
- Use the page attribute `page-edit-url-body` to provide placeholder
text in the body of the issue. The default text is`> Do not include
confidential information, personal data, sensitive data, or other
regulated data.`
- To add labels to the issue, use the page attribute
`page-edit-url-labels` (note that labels will be added only if they
already exist in the repo. New labels are not created.)