Fix canonicals (finally) (#300)
Requires pre-release versions to be marked as prerelease so Antora
doesn't consider them as 'latest'.

Pre-release versions will be included in the version selector, unless
`page-nav-selectors-hide-prerelease` is set to true.

This PR also modifies `page.canonicalUrl`. After it is set by Antora, a
helper in the ui-bundle checks the value, and if it contains the version
name that is the latest version of the page's component, the url is
updated by replacing the version name with 'current'. This ensures that
all our canonicals point to current (except where there is no equivalent
of a page in the latest release. In that case the canonical is
self-referencing).