Ensure only heading element text is used (#328)
This PR is required by a future change to how roles on headings are
handled.

Currently labels are processed by javascript when a page is loaded in a
browser. We are adding an extension to convert roles into labels during
the asciidoc to html conversion so that their metadata (dataset
attributes) are part of the raw HTML and can be used by tools that parse
the HTML.

This PR ensures that only the text content of the heading element is
used for the entries in the 'Contents' nav section of a page.