fix cheat sheet toc (#330)
A recent change to how the titles for Page Contents are generated
created a bug in the Cheat Sheet where the titles were not displayed in
the TOC, because the code was not correctly handling headings that
contain links. We don't typically use links in headings in the docs (in
fact it should possibly be a style guide recommendation to NOT use links
in headings in other docs) but the cheat sheet makes extensive use of
this to provide links to the relevant sections of the Cypher Manual.

This PR fixes the issue.