Fix google analytics bug (#344)
Fixes a critical issue where the version selector was unusable.

The issue is caused by a Google Analytics event that we have been
sending to track clicks in the version selector. This worked fine until
today, and may be related in some way to recent changes in the web
site build. It's very difficult to completely verify exact cause, though.